GOODBYE OLD OPMODES!
