+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/AutoWarehouse.java
@@ -1,5 +1,7 @@
 package org.firstinspires.ftc.teamcode.opmodes.auto;
 
+import static org.firstinspires.ftc.teamcode.opmodes.util.StayInPosition.stayInPose;
+import static org.firstinspires.ftc.teamcode.opmodes.util.VisionToLiftHeight.getPosition;
 import static org.firstinspires.ftc.teamcode.roadrunner.drive.RoadRunnerHelper.inchesToCoordinate;
 
 import com.acmerobotics.dashboard.telemetry.MultipleTelemetry;
@@ -7,29 +9,26 @@ import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+import com.qualcomm.robotcore.util.ElapsedTime;
 
-import org.firstinspires.ftc.teamcode.core.robot.tools.headless.AutoGrabber;
 import org.firstinspires.ftc.teamcode.core.robot.tools.headless.AutoIntake;
 import org.firstinspires.ftc.teamcode.core.robot.tools.headless.AutoLift;
 import org.firstinspires.ftc.teamcode.core.robot.vision.robot.TseDetector;
 import org.firstinspires.ftc.teamcode.core.thread.EventThread;
 import org.firstinspires.ftc.teamcode.core.thread.types.impl.TimedEvent;
-import org.firstinspires.ftc.teamcode.opmodes.util.PoseStorage;
+import org.firstinspires.ftc.teamcode.opmodes.util.WallSmash;
 import org.firstinspires.ftc.teamcode.roadrunner.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.roadrunner.trajectorysequence.TrajectorySequence;
-import org.firstinspires.ftc.teamcode.roadrunner.trajectorysequence.TrajectorySequenceBuilder;
-
-import java.util.Objects;
 
 @Autonomous
 public class AutoWarehouse extends LinearOpMode {
-    protected int multiplier = 1;
-    protected boolean isRed = false;
+    public int multiplier = 1;
+    public boolean isRed = false;
 
     @Override
     public void runOpMode() throws InterruptedException {
         TseDetector detector = new TseDetector(hardwareMap, "webcam", true, isRed);
-        final int[] height = {-1};
+        int height;
         double nextToWall = 70 - inchesToCoordinate(5.8D);
 
         MultipleTelemetry goodTelemetry = new MultipleTelemetry(telemetry);
@@ -43,98 +42,119 @@ public class AutoWarehouse extends LinearOpMode {
         final Pose2d initial = new Pose2d(0, multiplier * 70 - inchesToCoordinate(9),
                 Math.toRadians(90 * multiplier));
         drive.setPoseEstimate(initial);
+        final Pose2d liftPosition = new Pose2d(-2, 43.5 * multiplier,
+                Math.toRadians(85 * multiplier));
 
-        TrajectorySequence startSequence;
+        ElapsedTime toolTimer = new ElapsedTime();
+        ElapsedTime wallSmashTimer = new ElapsedTime();
 
-        {
-            TrajectorySequenceBuilder builder = drive.trajectorySequenceBuilder(initial);
-            builder.lineTo(new Vector2d(-3, 58 * multiplier));
-            // Go to lift
-            builder.lineToLinearHeading(new Pose2d(-2, 43.5 * multiplier,
-                    Math.toRadians(70 * multiplier)));
-            builder.addTemporalMarker(() -> {
-                lift.setPosition(getPosition(height[0]));
-            });
-            builder.waitSeconds(3.75);
-            builder.lineToLinearHeading(new Pose2d(0, (nextToWall + 1) * multiplier));
-            builder.addTemporalMarker(() -> drive.setWeightedDrivePower(new Pose2d(0, -0.2 * multiplier, 0)));
-            builder.lineTo(new Vector2d(20, (nextToWall + 1) * multiplier));
-            // go to warehouse
-            builder.lineTo(new Vector2d(40, (nextToWall + 1) * multiplier));
-            builder.addTemporalMarker(() -> drive.setWeightedDrivePower(new Pose2d(0, 0, 0)));
-            startSequence = builder.build();
-        }
+        // Part 1: drive to alliance shipping hub
+        final TrajectorySequence part1 = drive.trajectorySequenceBuilder(initial)
+                .lineTo(new Vector2d(-3, 58 * multiplier))
+                .lineToLinearHeading(liftPosition)
+                .build();
 
-        TrajectorySequence secondSequence;
+        // where the robot **should** be after you intake
+        final Pose2d intakeReturnPoint = new Pose2d(40, (nextToWall) * multiplier,
+                0);
 
-        {
-            TrajectorySequenceBuilder builder = drive.trajectorySequenceBuilder(new Pose2d(40,
-                    nextToWall * multiplier, Math.toRadians(0)));
-            builder.lineTo(new Vector2d(-3, nextToWall * multiplier));
-            builder.lineToLinearHeading(new Pose2d(-2, 43.5 * multiplier,
-                    Math.toRadians(70 * multiplier)));
-            builder.addTemporalMarker(() -> {
-                lift.setPosition(AutoLift.Positions.TOP);
-            });
-            builder.waitSeconds(4);
-            builder.lineToLinearHeading(new Pose2d(0, (nextToWall + 1) * multiplier));
-            builder.addTemporalMarker(() -> drive.setWeightedDrivePower(new Pose2d(0, -0.2 * multiplier, 0)));
-            builder.lineTo(new Vector2d(40, (nextToWall + 1) * multiplier));
-            builder.addTemporalMarker(() -> drive.setWeightedDrivePower(new Pose2d(0, 0, 0)));
+        // part 2: go to warehouse
+        final TrajectorySequence part2 = drive.trajectorySequenceBuilder(liftPosition)
+                .lineToLinearHeading(new Pose2d(0, (nextToWall) * multiplier))
+                .addDisplacementMarker(() -> drive.setWeightedDrivePower(new Pose2d(0, -0.2 * multiplier, 0)))
+                .lineTo(intakeReturnPoint.vec())
+                .build();
 
-            secondSequence = builder.build();
-        }
 
-        Thread thread = new Thread(() -> {
+        // part 3: move back to Alliance Shipping hub. then you can go back to part 2 as needed.
+        final TrajectorySequence part3 = drive.trajectorySequenceBuilder(intakeReturnPoint)
+                .lineTo(new Vector2d(-3, nextToWall * multiplier))
+                .lineToLinearHeading(liftPosition)
+                .build();
+
+        final boolean[] liftUpdated = {false};
+        Thread liftThread = new Thread(() -> {
             while (!isStopRequested()) {
                 lift.update();
+                liftUpdated[0] = true;
             }
         });
 
         waitForStart();
-        height[0] = detector.run();
-        goodTelemetry.addData("height", height[0]);
-        goodTelemetry.update();
-
-        thread.start();
+        liftThread.start();
         eventThread.start();
 
-        if (!isStopRequested()) {
+        intake.lightsOff();
+        height = detector.run();
+        intake.lightsOn();
+        goodTelemetry.addData("height", height);
+        goodTelemetry.update();
 
-            drive.followTrajectorySequence(startSequence);
-
-            for (int i = 0; i < 1; i++) {
-                // intake a block
-                intake.backward();
-                drive.setWeightedDrivePower(new Pose2d(0.2, 0, 0));
-                while (intake.noObject()) {
-                    if (isStopRequested()) {
-                        return;
-                    }
-                }
-                eventThread.addEvent(new TimedEvent(intake::stop, 250));
-                drive.setWeightedDrivePower(new Pose2d(0, 0, 0));
-                drive.followTrajectory(
-                        drive.trajectoryBuilder(drive.getPoseEstimate())
-                                .lineToLinearHeading(new Pose2d(40, nextToWall + 2 * multiplier,
-                                        Math.toRadians(0)))
-                                .build()
-                );
-
-                {
-                    Pose2d pose = drive.getPoseEstimate();
-                    drive.setPoseEstimate(new Pose2d(pose.getX(), nextToWall * multiplier, pose.getHeading()));
-                }
-
-                drive.followTrajectorySequence(secondSequence);
+        drive.followTrajectorySequenceAsync(part1);
+        updateLoop(drive);
+        liftUpdated[0] = false;
+        lift.setPosition(getPosition(height));
+        toolTimer.reset();
+        while (toolTimer.seconds() < 3) {
+            if (isStopRequested()) {
+                return;
             }
+            stayInPose(drive, part1.end());
         }
-        drive.followTrajectorySequenceAsync(null);
-        PoseStorage.currentPose = drive.getPoseEstimate();
+
+        drive.followTrajectorySequenceAsync(part2);
+        updateLoop(drive);
+        if (isStopRequested()) return;
+
+        WallSmash.smashIntoWall(drive, multiplier, 500);
+
+        intake(toolTimer, drive, intake);
+
+        WallSmash.smashIntoWall(drive, multiplier, 500);
+
+        drive.followTrajectorySequenceAsync(part3);
+        updateLoop(drive);
+        if (isStopRequested()) return;
+
+        lift.setPosition(AutoLift.Positions.TOP);
+        toolTimer.reset();
+        while (toolTimer.seconds() < 3) {
+            if (isStopRequested()) {
+                return;
+            }
+            stayInPose(drive, part3.end());
+        }
+
+        drive.followTrajectorySequenceAsync(part2);
+        updateLoop(drive);
+        intake(toolTimer, drive, intake);
+        while (!isStopRequested()) {
+            stayInPose(drive, part2.end());
+        }
     }
 
-    public AutoLift.Positions getPosition(int input) {
-        return input == 1 ? AutoLift.Positions.BOTTOM :
-                input == 2 ? AutoLift.Positions.MIDDLE : AutoLift.Positions.TOP;
+    public void intake(ElapsedTime timer, SampleMecanumDrive drive, AutoIntake intake) {
+        // intake a block
+        intake.backward();
+        drive.setWeightedDrivePower(new Pose2d(0.2, 0, 0));
+        while (intake.noObject()) {
+            if (isStopRequested()) {
+                return;
+            }
+        }
+        timer.reset();
+        while (timer.milliseconds() < 250) {}
+        intake.stop();
+        intake.forward();
+        timer.reset();
+        while (timer.milliseconds() < 100) {}
+        intake.stop();
+        drive.setWeightedDrivePower(new Pose2d(0, 0, 0));
+    }
+
+    public void updateLoop(SampleMecanumDrive drive) {
+        while (!isStopRequested() && drive.isBusy()) {
+            drive.update();
+        }
     }
 }
