--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/RedNewAutoStorage.java
@@ -1,11 +1,0 @@
-package org.firstinspires.ftc.teamcode.opmodes.auto;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-
-@Autonomous
-public class RedNewAutoStorage extends NewAutoStorage {
-    public RedNewAutoStorage() {
-        isRed = true;
-        multiplier = -1;
-    }
-}
