--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/RedNewAutoWarehouse.java
@@ -1,11 +1,0 @@
-package org.firstinspires.ftc.teamcode.opmodes.auto;
-
-import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-
-@Autonomous
-public class RedNewAutoWarehouse extends NewAutoWarehouse {
-    public RedNewAutoWarehouse() {
-        isRed = true;
-        multiplier = -1;
-    }
-}
